Enable speed search for and clean up popup
