+++ src/main/kotlin/com/fwdekker/randomness/PopupAction.kt
@@ -3,7 +3,6 @@ package com.fwdekker.randomness
 import com.fwdekker.randomness.template.TemplateGroupAction
 import com.fwdekker.randomness.template.TemplateSettings
 import com.fwdekker.randomness.template.TemplateSettingsAction
-import com.fwdekker.randomness.ui.registerModifierActions
 import com.intellij.openapi.actionSystem.ActionGroup
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -12,6 +11,10 @@ import com.intellij.openapi.actionSystem.Separator
 import com.intellij.openapi.ui.popup.JBPopupFactory
 import com.intellij.ui.popup.list.ListPopupImpl
 import java.awt.event.ActionEvent
+import java.awt.event.KeyEvent
+import java.util.Locale
+import javax.swing.AbstractAction
+import javax.swing.KeyStroke
 
 
 /**
@@ -49,7 +52,6 @@ class PopupAction : AnAction() {
             )
             as ListPopupImpl
 
-        popup.speedSearch.setEnabled(false)
         if (hasEditor) {
             popup.setCaption(Bundle("popup.title"))
             popup.setAdText(Bundle("popup.ad"))
@@ -118,5 +120,105 @@ class PopupAction : AnAction() {
             TemplateSettings.default.state.templates.map { TemplateSettingsAction(it) }.toTypedArray<AnAction>() +
                 Separator() +
                 TemplateSettingsAction()
+    }
+}
+
+
+/**
+ * An `AbstractAction` that uses [myActionPerformed] as the implementation of its [actionPerformed] method.
+ *
+ * @property myActionPerformed The code to execute in [actionPerformed].
+ */
+private class SimpleAbstractAction(private val myActionPerformed: (ActionEvent?) -> Unit) : AbstractAction() {
+    override fun actionPerformed(event: ActionEvent?) = myActionPerformed(event)
+}
+
+/**
+ * Returns the cartesian product of two lists.
+ *
+ * By requiring both lists to actually be lists of lists, this method can be chained.
+ *
+ * Consider the following examples, using a simplified notation for lists for readability:
+ * ```
+ * $ [[1, 2]] * [[3, 4]]
+ * [[1, 3], [1, 4], [2, 3], [2, 4]]
+ *
+ * $ [[1, 2]] * [[3, 4]] * [[5, 6]]
+ * [[1, 3, 5], [1, 3, 6], [1, 4, 5], [1, 4, 6], [2, 3, 5], [2, 3, 6], [2, 4, 5], [2, 4, 6]]
+ * ```
+ *
+ * @param E the type of inner element
+ * @param other the list to multiply with
+ * @return the cartesian product of `this` and [other]
+ */
+@Suppress("UnusedPrivateMember") // False positive: Used as operator `*`
+private operator fun <E> List<List<E>>.times(other: List<List<E>>) =
+    this.flatMap { t1 -> other.map { t2 -> t1 + t2 } }
+
+/**
+ * Registers actions such that actions can be selected while holding (combinations of) modifier keys.
+ *
+ * All combinations of modifier keys are registered for events. Additionally, the [captionModifier] function is invoked
+ * every time the user presses or releases any modifier key, even while holding other modifier keys.
+ *
+ * Events are also registered for pressing the Enter key (with or without modifier keys) to invoke the action that is
+ * currently highlighted, and for pressing one of the numbers 1-9 (with or without modifier keys) to invoke the action
+ * at that index in the popup.
+ *
+ * @param captionModifier returns the caption to set based on the event
+ */
+fun ListPopupImpl.registerModifierActions(captionModifier: (ActionEvent?) -> String) {
+    val modifiers = listOf(listOf("alt"), listOf("control"), listOf("shift"))
+    val optionalModifiers = listOf(listOf("")) + modifiers
+
+    (modifiers * optionalModifiers * optionalModifiers).forEach { (a, b, c) ->
+        registerAction(
+            "${a}${b}${c}Released",
+            KeyStroke.getKeyStroke("$b $c released ${a.uppercase(Locale.getDefault())}"),
+            SimpleAbstractAction { setCaption(captionModifier(it)) }
+        )
+        registerAction(
+            "${a}${b}${c}Pressed",
+            KeyStroke.getKeyStroke("$a $b $c pressed ${a.uppercase(Locale.getDefault())}"),
+            SimpleAbstractAction { setCaption(captionModifier(it)) }
+        )
+        registerAction(
+            "${a}${b}${c}invokeAction",
+            KeyStroke.getKeyStroke("$a $b $c pressed ENTER"),
+            SimpleAbstractAction { event ->
+                event ?: return@SimpleAbstractAction
+
+                handleSelect(
+                    true,
+                    KeyEvent(
+                        component,
+                        event.id, event.getWhen(), event.modifiers,
+                        KeyEvent.VK_ENTER, KeyEvent.CHAR_UNDEFINED, KeyEvent.KEY_LOCATION_UNKNOWN
+                    )
+                )
+            }
+        )
+
+        @Suppress("MagicNumber") // Not worth a constant
+        for (key in 0..9) {
+            registerAction(
+                "${a}${b}${c}invokeAction$key",
+                KeyStroke.getKeyStroke("$a $b $c pressed $key"),
+                SimpleAbstractAction { event ->
+                    event ?: return@SimpleAbstractAction
+
+                    val targetRow = if (key == 0) 9 else key - 1
+                    list.addSelectionInterval(targetRow, targetRow)
+                    handleSelect(
+                        true,
+                        KeyEvent(
+                            component,
+                            event.id, event.getWhen(), event.modifiers,
+                            KeyEvent.VK_ENTER, KeyEvent.CHAR_UNDEFINED, KeyEvent.KEY_LOCATION_UNKNOWN
+                        )
+                    )
+                }
+            )
+        }
     }
 }
